requirements.txt
djoser
Django
djangorestframework==3.12.4
pillow
qrcode
djangorestframework-simplejwt==4.7.2
django-rest-auth
pytz
gunicorn
psycopg2-binary
django-phonenumber-field[phonenumberslite]
django-allauth
webcolors